mediump -> highp (w/notes) to try and fix iphone6(+) hail pixel issues

+++ cocos/renderer/ccShader_3D_Terrain.frag
@@ -1,7 +1,8 @@
 
 const char* cc3D_Terrain_frag = R"(
 #ifdef GL_ES
-precision mediump float;
+// STEVE: precision mediump float;
+precision highp float;
 #endif
 uniform vec3 u_color;
 varying vec2 v_texCoord;
