+++ cocos/renderer/ccShader_Label_normal.frag
@@ -25,7 +25,8 @@
 
 const char* ccLabelNormal_frag = R"(
 #ifdef GL_ES
-precision lowp float;
+// STEVE: precision lowp float;
+precision highp float;
 #endif
 
 varying vec4 v_fragmentColor;
