+++ cocos/renderer/ccShader_Position.vert
@@ -27,7 +27,8 @@ const char* ccPosition_vert = R"(
 attribute vec4 a_position;
 
 #ifdef GL_ES
-varying lowp vec4 v_position;
+// STEVE: varying lowp vec4 v_position;
+varying highp vec4 v_position;
 #else
 varying vec4 v_position;
 #endif
