+++ cocos/renderer/ccShader_PositionTexture.frag
@@ -26,7 +26,8 @@
 const char* ccPositionTexture_frag = R"(
 
 #ifdef GL_ES
-precision lowp float;
+// STEVE: precision lowp float;
+precision highp float;
 #endif
 
 varying vec2 v_texCoord;
