+++ cocos/renderer/ccShader_PositionTextureColor_noMVP.vert
@@ -29,8 +29,10 @@ attribute vec2 a_texCoord;
 attribute vec4 a_color;
 
 #ifdef GL_ES
-varying lowp vec4 v_fragmentColor;
-varying mediump vec2 v_texCoord;
+// STEVE: varying lowp vec4 v_fragmentColor;
+varying mediump vec4 v_fragmentColor;
+// STEVE: varying mediump vec2 v_texCoord;
+varying highp vec2 v_texCoord;
 #else
 varying vec4 v_fragmentColor;
 varying vec2 v_texCoord;
