+++ cocos/renderer/ccShader_Position_uColor.vert
@@ -30,7 +30,8 @@ uniform vec4 u_color;
 uniform float u_pointSize;
 
 #ifdef GL_ES
-varying lowp vec4 v_fragmentColor;
+// STEVE: varying lowp vec4 v_fragmentColor;
+varying highp vec4 v_fragmentColor;
 #else
 varying vec4 v_fragmentColor;
 #endif
